requirements.txt
# Web Framework
streamlit==1.39.0

# Backend Services
supabase==2.14.0

# AI/ML
google-generativeai==0.8.3

# Image Processing
pillow==10.4.0
qrcode[pil]==8.0

# UI Components
streamlit-autorefresh==1.0.1

# Utilities
python-dotenv==1.0.1